feat: Add CreateUser entity
This commit adds the CreateUser entity to the user package. The CreateUser entity represents the information needed to create a new user, including the username, password, email, userType, and userDetails. This entity will be used in the process of creating new users in the system.

Refactor the code to include the CreateUser entity and its associated methods and interfaces.